Add fallback when no plugin returns a value on getLineHTMLForExport()
Etherpad only exports context.lineContent on HTML export when at least
one plugin returns a value. If this is the only plugin installed, or if
none of the installed plugins retuns a value, all code made on
exportHTML.js will be ignored. To fix this, it was necessary to add a
fallback style on exported CSS.